requirements.txt
absl-py @ file:///C:/ci/absl-py_1603893176799/work
aiml==0.9.2
aiohttp @ file:///C:/ci/aiohttp_1602530358780/work
astor==0.8.1
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
azure-cognitiveservices-language-textanalytics==0.2.0
azure-common==1.1.27
beautifulsoup4==4.9.3
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1600699250966/work
chardet==3.0.4
click==7.1.2
cryptography @ file:///C:/ci/cryptography_1601046913206/work
cycler==0.10.0
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1604100403705/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1597406403308/work
h5py==3.1.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
isodate==0.6.0
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
Keras==2.3.1
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
kiwisolver @ file:///C:/ci/kiwisolver_1603996595157/work
lxml==4.6.1
Markdown @ file:///C:/ci/markdown_1603216569530/work
mask-rcnn==2.1
matplotlib @ file:///C:/ci/matplotlib-base_1603356257853/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
msrest==0.6.21
multidict @ file:///C:/ci/multidict_1600456486794/work
multitasking==0.0.9
nltk @ file:///tmp/build/80754af9/nltk_1592496090529/work
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
pandas==1.1.4
Pillow @ file:///C:/ci/pillow_1609786872067/work
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
regex @ file:///C:/ci/regex_1602782614108/work
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
scikit-learn @ file:///C:/ci/scikit-learn_1598376983131/work
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
Werkzeug==0.16.1
wikipedia==1.4.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yahoo-finance==1.4.0
yarl @ file:///C:/ci/yarl_1602679793616/work
yfinance==0.1.55
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
